Updated Feature Flow
Separated the flow demo in file as that will make the standard way of making features.
