+++ app/src/main/java/com/example/jetpack_compose_all_in_one/features/flows_demo/StateFlowVsSharedFlow.kt
@@ -7,58 +7,21 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringArrayResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.example.jetpack_compose_all_in_one.R
-import com.example.jetpack_compose_all_in_one.ui.components.LessonHeader
 import com.example.jetpack_compose_all_in_one.ui.components.SimpleTextButton
 import com.example.jetpack_compose_all_in_one.ui.theme.dp_15
 import com.example.jetpack_compose_all_in_one.ui.theme.dp_30
-import com.example.jetpack_compose_all_in_one.utils.LogicPager
 
 @Composable
-fun FeatureFlowContent() {
-    val currentPage = rememberSaveable { mutableIntStateOf(0) }
-
-    LogicPager(
-        pageCount = 2,
-        currentPage = currentPage
-    ) {
-        Column(
-            Modifier
-                .fillMaxSize()
-                .padding(it)
-        ) {
-            LessonHeader(
-                stringArrayResource(R.array.feature_flow_example)[currentPage.value],
-                Modifier
-                    .fillMaxWidth()
-                    .padding(dp_15),
-                TextAlign.Center
-            )
-
-            when (currentPage.value) {
-                0 -> FlowDemoWithMultipleConsumer()
-                1 -> StateFlowVsSharedFlow()
-            }
-        }
-    }
-}
-
-@Composable
-private fun StateFlowVsSharedFlow() {
+fun StateFlowVsSharedFlow() {
     val flowsViewModel: FlowsViewModel = hiltViewModel()
     val state1 by flowsViewModel.stateFlow1.collectAsState()
     val state2 by flowsViewModel.sharedFlow1.collectAsState(0)
